# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Desperately Seeking - buyer-first marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get the currently active user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateCurrentUser
      tags: [users]
      summary: Update profile of currently active user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /users/switch:
    post:
      operationId: switchUser
      tags: [users]
      summary: Switch the active user (demo)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId:
                  type: string
      responses:
        "200":
          description: New current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user profile by id
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
  /requests/{requestId}/responses:
    get:
      operationId: listResponsesForRequest
      tags: [responses]
      summary: List seller responses for a request
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Responses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SellerResponse"
    post:
      operationId: createResponse
      tags: [responses]
      summary: Submit a seller response with photos and price
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResponseBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerResponse"
  /responses/{responseId}:
    get:
      operationId: getResponse
      tags: [responses]
      summary: Get a response by id
      parameters:
        - in: path
          name: responseId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerResponse"
    patch:
      operationId: updateResponseStatus
      tags: [responses]
      summary: Buyer accepts or declines a seller response
      parameters:
        - in: path
          name: responseId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, accepted, declined]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellerResponse"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, handle, avatarUrl, joinedAt, bio, location]
      properties:
        id:
          type: string
        name:
          type: string
        handle:
          type: string
        avatarUrl:
          type: string
        joinedAt:
          type: string
          format: date-time
        bio:
          type: string
        location:
          type: string
    UserProfile:
      type: object
      required: [user, requestCount, responseCount, acceptedCount]
      properties:
        user:
          $ref: "#/components/schemas/User"
        requestCount:
          type: integer
        responseCount:
          type: integer
        acceptedCount:
          type: integer
        recentRequests:
          type: array
          items:
            $ref: "#/components/schemas/RequestSummary"
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        location:
          type: string
        avatarUrl:
          type: string
    RequestSummary:
      type: object
      required:
        [
          id,
          title,
          description,
          category,
          status,
          urgency,
          createdAt,
          buyer,
          responseCount,
          location,
        ]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        budgetMin:
          type: number
          nullable: true
        budgetMax:
          type: number
          nullable: true
        status:
          type: string
          enum: [open, fulfilled, closed]
        urgency:
          type: string
          enum: [low, normal, high]
        location:
          type: string
        createdAt:
          type: string
          format: date-time
        buyer:
          $ref: "#/components/schemas/User"
        responseCount:
          type: integer
...
    SellerResponse:
      type: object
      required:
        [
          id,
          requestId,
          seller,
          price,
          condition,
          message,
          photos,
          status,
          createdAt,
        ]
      properties:
        id:
          type: string
        requestId:
          type: string
        seller:
          $ref: "#/components/schemas/User"
        price:
          type: number
        condition:
          type: string
          enum: [new, like_new, good, fair, used]
        message:
          type: string
        photos:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, accepted, declined]
        threadId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateResponseBody:
      type: object
      required: [price, condition, message, photos]
      properties:
        price:
          type: number
        condition:
          type: string
          enum: [new, like_new, good, fair, used]
        message:
          type: string
          minLength: 1
        photos:
          type: array
          items:
            type: string
...