# lib/api-zod/src/generated/types/createResponseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Desperately Seeking - buyer-first marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateResponseBodyCondition } from "./createResponseBodyCondition";

export interface CreateResponseBody {
  price: number;
  condition: CreateResponseBodyCondition;
  /** @minLength 1 */
  message: string;
  photos: string[];
}